컨퍼런스,세션 리팩토링
