--- src/main/java/goorm/back/zo6/conference/application/ConferenceCreateService.java
@@ -1,5 +1,0 @@
-package goorm.back.zo6.conference.application;
-
-public interface ConferenceCreateService {
-    ConferenceResponse createConference(ConferenceCreateRequest request);
-}
