--- src/main/java/goorm/back/zo6/conference/application/ConferenceSimpleResponse.java
@@ -1,27 +1,0 @@
-package goorm.back.zo6.conference.application;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.format.annotation.DateTimeFormat;
-
-import java.time.LocalDateTime;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ConferenceSimpleResponse {
-    private Long conferenceId;
-
-    private String conferenceName;
-
-    @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    private LocalDateTime startTime;
-
-    @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    private LocalDateTime endTime;
-
-    private String conferenceImageUrl;
-
-    private String conferenceLocation;
-}
