--- src/main/java/goorm/back/zo6/conference/application/SessionCreateService.java
@@ -1,5 +1,0 @@
-package goorm.back.zo6.conference.application;
-
-public interface SessionCreateService {
-    SessionResponse createSession(SessionCreateRequest request);
-}
