--- src/main/java/goorm/back/zo6/conference/application/SessionService.java
@@ -1,5 +1,0 @@
-package goorm.back.zo6.conference.application;
-
-public interface SessionService {
-    SessionResponse updateSession(Long sessionId, String location);
-}
