+++ src/main/java/goorm/back/zo6/conference/application/command/session/SessionCommandServiceImpl.java
@@ -1,0 +1,71 @@
+package goorm.back.zo6.conference.application.command.session;
+
+import goorm.back.zo6.conference.application.dto.SessionCreateRequest;
+import goorm.back.zo6.conference.application.dto.SessionDto;
+import goorm.back.zo6.conference.application.shared.ConferenceValidator;
+import goorm.back.zo6.conference.application.shared.SessionValidator;
+import goorm.back.zo6.conference.domain.Conference;
+import goorm.back.zo6.conference.domain.Session;
+import goorm.back.zo6.conference.domain.SessionRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class SessionCommandServiceImpl implements SessionCommandService {
+
+    private final SessionRepository sessionRepository;
+
+    private final ConferenceValidator conferenceValidator;
+
+    private final SessionValidator sessionValidator;
+
+    @Override
+    public boolean getSessionStatus(Long conferenceId, Long sessionId) {
+        Session session = sessionValidator.getSessionOrThrow(sessionId);
+        return session.isActive();
+    }
+
+    @Override
+    @Transactional
+    public SessionDto updateSession(Long sessionId, String location) {
+        Session session = sessionValidator.getSessionOrThrow(sessionId);
+        session.updateSession(location);
+
+        sessionRepository.save(session);
+
+        return SessionDto.fromEntity(session);
+    }
+
+    @Override
+    public void updateSessionStatus(Long conferenceId, Long sessionId, boolean newStatus) {
+        Session session = sessionValidator.findSessionByConferenceAndSessionId(conferenceId, sessionId);
+        session.updateActive(newStatus);
+        sessionRepository.save(session);
+    }
+
+    @Override
+    @Transactional
+    public SessionDto createSession(SessionCreateRequest request) {
+        Conference conference = conferenceValidator.findConferenceOrThrow(request.getConferenceId());
+
+        Session session = Session.builder()
+                .name(request.getName())
+                .capacity(request.getCapacity())
+                .location(request.getLocation())
+                .startTime(request.getStartTime())
+                .endTime(request.getEndTime())
+                .summary(request.getSummary())
+                .speakerName(request.getSpeakerName())
+                .speakerOrganization(request.getSpeakerOrganization())
+                .isActive(true)
+                .speakerImageKey(request.getSpeakerImage())
+                .conference(conference)
+                .build();
+
+        Session savedSession = sessionRepository.save(session);
+
+        return SessionDto.fromEntity(savedSession);
+    }
+}
