+++ src/main/java/goorm/back/zo6/conference/application/dto/ConferenceCreateRequest.java
@@ -1,16 +1,16 @@
-package goorm.back.zo6.conference.application;
+package goorm.back.zo6.conference.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.time.LocalDateTime;
 
 @Getter
 @AllArgsConstructor
-public class ConferenceResponse {
-
-    private Long id;
+@NoArgsConstructor
+public class ConferenceCreateRequest {
 
     private String name;
 
@@ -18,13 +18,13 @@ public class ConferenceResponse {
 
     private String location;
 
+    private Integer capacity;
+
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime startTime;
 
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime endTime;
-
-    private Integer capacity;
 
     private String imageUrl;
