+++ src/main/java/goorm/back/zo6/conference/application/query/ConferenceQueryService.java
@@ -1,0 +1,18 @@
+package goorm.back.zo6.conference.application.query;
+
+import goorm.back.zo6.conference.application.dto.ConferenceDetailResponse;
+import goorm.back.zo6.conference.application.dto.ConferenceResponse;
+import goorm.back.zo6.conference.application.dto.SessionDto;
+
+import java.util.List;
+
+public interface ConferenceQueryService {
+
+    List<ConferenceResponse> getAllConferences();
+
+    ConferenceDetailResponse getConference(Long conferenceId);
+
+    List<SessionDto> getSessionsByConferenceId(Long conferenceId);
+
+    SessionDto getSessionDetail(Long conferenceId, Long sessionId);
+}
