+++ src/main/java/goorm/back/zo6/conference/domain/Conference.java
@@ -10,7 +10,6 @@ import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -72,7 +71,7 @@ public class Conference {
     }
 
     public void validateSessionOwnership(Set<Long> sessionIds) {
-        if (!hasSessions) {
+        if (Boolean.FALSE.equals(hasSessions)) {
             throw new CustomException(ErrorCode.CONFERENCE_NOT_FOUND);
         }
