+++ src/main/java/goorm/back/zo6/conference/infrastructure/SessionRepositoryImpl.java
@@ -5,7 +5,6 @@ import goorm.back.zo6.conference.domain.SessionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
