+++ src/main/java/goorm/back/zo6/conference/presentation/ConferenceAdminController.java
@@ -1,7 +1,9 @@
 package goorm.back.zo6.conference.presentation;
 
-import goorm.back.zo6.conference.application.*;
-import goorm.back.zo6.conference.domain.Session;
+import goorm.back.zo6.conference.application.command.conference.ConferenceCommandService;
+import goorm.back.zo6.conference.application.command.session.SessionCommandService;
+import goorm.back.zo6.conference.application.dto.*;
+import goorm.back.zo6.conference.application.query.ConferenceQueryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -18,8 +20,12 @@ import java.util.List;
 @RequestMapping("/api/v1/admin/conference")
 @RequiredArgsConstructor
 public class ConferenceAdminController {
+
     private final ConferenceQueryService conferenceQueryService;
-    private final SessionService sessionService;
+
+    private final ConferenceCommandService conferenceCommandService;
+
+    private final SessionCommandService sessionCommandService;
 
     @GetMapping
     @Operation(summary = "모든 컨퍼런스 조회", description = "등록된 모든 컨퍼런스 목록을 조회합니다. (관리자 전용)")
@@ -38,33 +44,33 @@ public class ConferenceAdminController {
     @GetMapping("/{conferenceId}/sessions/{sessionId}")
     @Operation(summary = "특정 세션 상세 조회", description = "conferenceId와 sessionId로 특정 세션의 상세 정보를 조회합니다. (관리자 전용)")
     public ResponseEntity<SessionDto> getSessionDetail(@PathVariable Long conferenceId, @PathVariable Long sessionId) {
-        Session session = conferenceQueryService.getSessionDetail(conferenceId, sessionId);
-        return ResponseEntity.ok(SessionDto.fromEntity(session));
+        SessionDto session = conferenceQueryService.getSessionDetail(conferenceId, sessionId);
+        return ResponseEntity.ok(session);
     }
 
     @PutMapping("/{conferenceId}/status")
     @Operation(summary = "컨퍼런스 상태 변경", description = "활성화 / 비활성화를 변경합니다. (관리자 전용)")
     @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<String> conferenceStatus(@PathVariable Long conferenceId) {
-        boolean currentStatus = conferenceQueryService.getConferenceStatus(conferenceId);
+        boolean currentStatus = conferenceCommandService.getConferenceStatus(conferenceId);
         boolean newStatus = !currentStatus;
-        conferenceQueryService.updateConferenceStatus(conferenceId, newStatus);
+        conferenceCommandService.updateConferenceStatus(conferenceId, newStatus);
         return ResponseEntity.ok("상태가 변경되었습니다.");
     }
 
     @PutMapping("/{conferenceId}/sessions/{sessionId}")
     @Operation(summary = "세션 상태 변경", description = "활성화 / 비활성화를 변경합니다. (관리자 전용)")
     public ResponseEntity<String> sessionStatus(@PathVariable Long conferenceId, @PathVariable Long sessionId) {
-        boolean currentStatus = conferenceQueryService.getSessionStatus(conferenceId, sessionId);
+        boolean currentStatus = sessionCommandService.getSessionStatus(conferenceId, sessionId);
         boolean newStatus = !currentStatus;
-        conferenceQueryService.updateSessionStatus(conferenceId, sessionId, newStatus);
+        sessionCommandService.updateSessionStatus(conferenceId, sessionId, newStatus);
         return ResponseEntity.ok("상태가 변경되었습니다.");
     }
 
     @PutMapping("/sessions/{sessionId}")
     @Operation(summary = "세션 정보 수정", description = "세션의 정보를 수정합니다 (구역만 수정 가능).")
-    public ResponseEntity<SessionResponse> updateSession(@PathVariable Long sessionId, @Valid @RequestBody SessionUpdateRequest request) {
-        SessionResponse response = sessionService.updateSession(
+    public ResponseEntity<SessionDto> updateSession(@PathVariable Long sessionId, @Valid @RequestBody SessionUpdateRequest request) {
+        SessionDto response = sessionCommandService.updateSession(
                 sessionId,
                 request.getLocation()
         );
