+++ src/main/java/goorm/back/zo6/conference/presentation/ConferenceController.java
@@ -1,7 +1,9 @@
 package goorm.back.zo6.conference.presentation;
 
-import goorm.back.zo6.conference.application.*;
-import goorm.back.zo6.conference.domain.Session;
+import goorm.back.zo6.conference.application.dto.ConferenceDetailResponse;
+import goorm.back.zo6.conference.application.dto.ConferenceResponse;
+import goorm.back.zo6.conference.application.dto.SessionDto;
+import goorm.back.zo6.conference.application.query.ConferenceQueryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -35,7 +37,7 @@ public class ConferenceController {
     @GetMapping("/{conferenceId}/sessions/{sessionId}")
     @Operation(summary = "특정 세션 상세 조회", description = "conferenceId와 sessionId로 특정 세션의 상세 정보를 조회합니다.")
     public ResponseEntity<SessionDto> getSessionDetail(@PathVariable Long conferenceId, @PathVariable Long sessionId) {
-        Session session = conferenceQueryService.getSessionDetail(conferenceId, sessionId);
-        return ResponseEntity.ok(SessionDto.fromEntity(session));
+        SessionDto session = conferenceQueryService.getSessionDetail(conferenceId, sessionId);
+        return ResponseEntity.ok(session);
     }
 }
