+++ src/main/java/goorm/back/zo6/conference/presentation/ConferenceCreateController.java
@@ -1,9 +1,8 @@
 package goorm.back.zo6.conference.presentation;
 
-import goorm.back.zo6.conference.application.ConferenceCreateRequest;
-import goorm.back.zo6.conference.application.ConferenceCreateService;
-import goorm.back.zo6.conference.application.ConferenceResponse;
-import goorm.back.zo6.conference.domain.Conference;
+import goorm.back.zo6.conference.application.command.conference.ConferenceCommandService;
+import goorm.back.zo6.conference.application.dto.ConferenceCreateRequest;
+import goorm.back.zo6.conference.application.dto.ConferenceResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -15,19 +14,18 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-
 @Tag(name = "conference", description = "Conference API")
 @RestController
 @RequestMapping("/conferences")
 @RequiredArgsConstructor
 public class ConferenceCreateController {
 
-    private final ConferenceCreateService conferenceCreateService;
+    private final ConferenceCommandService conferenceCommandService;
 
     @PostMapping
     @Operation(summary = "컨퍼런스 생성", description = "새로운 컨퍼런스를 생성합니다.")
     public ResponseEntity<ConferenceResponse> createConference(@Valid @RequestBody ConferenceCreateRequest request) {
-        ConferenceResponse conferenceResponse = conferenceCreateService.createConference(request);
+        ConferenceResponse conferenceResponse = conferenceCommandService.createConference(request);
 
         return ResponseEntity.status(HttpStatus.CREATED).body(conferenceResponse);
     }
